The first commit of a series of 6 commits for the amd64 port.
This commit includes header files for amd64 cpu that are in
sys/cpu/amd64/include.

The #1 task to port for amd64 is to get the world/kernel cross-compiled
for the 64-bit target. This involves creating a platform and cpu target
for the build.  The skeleton infrastructure for the cpu and platform is
the sys/cpu/amd64 directory and the sys/platform/pc64 directory.

The first stage of this #1 task is to have the world cross-build.  It
includes the building of cross toolchains, the libraries and the world.
This series of commits makes the world cross-build (buildworld) for the
amd64 port with these environment variables:
TARGET_CPU=amd64; TARGET_PLATFORM=pc64; NO_GCC41=1

This is the first commit: resurrect headers for sys/cpu/amd64/include from
CVS Attic; patch and mark them as CPU specific.

On-behalf-of:  Noah Yan <noah.yan@gmail.com>
Submitted-by:  Noah Yan <noah.yan@gmail.com>