Rework generating of files so that every generation tool will only run once.
Before, we generated a file over and over again where it was needed by the
build process, instead of generating the files once and reusing them.
This commit changes this behavior and makes the build match more gcc's own
build way, which is quite optimized.

Additionally, use move-if-change, to avoid rebuilds if the generated files
didn't change.