Fix a bug in linprocfs and hold lwps to prevent a race condition.
Linprocfs rolls its own prototype for procfs_domem.  The function was
changed to take a struct lwp *, but neither the private prototype, nor
the procfs_domem call were changed accordingly.  As a result any access
to the linprocfs process memory file resulted in a panic.

Procfs and linprocs perform potentially blocking calls, so HOLD and
RELE the lwps so that they can't be destroyed while procfs is blocked.