Sync with FreeBSD.
Make it possible to have this (basename(1)) perform basename(3) on multiple
files.  The traditional behaviour, 'basename string .suffix', is preserved,
however a suffix may now also be specified via a getopt(3) option, -s, such
that if it is specified in that way, all string arguments follow.  There is
also a new flag, -a, which allows a user to say "yes, please basename(3) on
all arguments".

Add EXIT STATUS.

Cross-reference basename(3) and dirname(3).

Obtained-from:  FreeBSD