Adjust the description of HAMMER's storage limitations.  I have rearranged the way clusters are organized in order to implement a layered storage management mechanic within the cluster.  To do this the number of 16K filesystem buffers in a cluster had to be limited to 4096.  This allows several layered A-lists to be embedded in the cluster header instead of just one but limits the cluster size to 64MB instead of 512M.  Without any other changes this would have limited volumes to 2TB, which is too small.
Add a new super-cluster layer to support volumes larger then 2TB.  Instead
of being limited to 32768 clusters per volume we now have a limit of
16384 super-clusters per volume and 32768 clusters per super-cluster,
resulting in a limit of 32768TB per volume.

A super-cluster is just another 16K filesystem buffer which adds another
A-List layer to the volume's storage management scheme.  It turns out to
be totally trivial to implement.  Cluster allocations will still be optimal
in a super-cluster arrangement simply by caching the last super-cluster
an allocation was made from.