Clean up the structural organization.  Separate out A-lists and make the header filenames conform to a single standard.  Add a super-cluster header and reorganize the A-list topology used in the cluster header.
Augment A-lists to support hinting across A-list layers.  This allows
A-lists to be glued together hierarchically and to pass hinting information
between the layers.  This support will be used by HAMMER to manage cluster
allocation via the A-lists embedded in volume and super-cluster headers,
and to manage B-Tree, record, and data allocation via the A-lists embedded
in the cluster header and embedded in individual filesystem buffers.