HAMMER 9/many - btree removal cases, mount nohistory
Add a 'nohistory' mount option that will cause HAMMER to not retain any
history.  This option is primarily for testing of btree removal and hinted
radix tree bitmap frees and reallocations.

Flesh out the btree node removal code.  We don't try to rebalance the tree
yet but we do attempt to remove empty nodes.

Add workarounds for a GCC-4 bug involving overflow tests on integers.