Call selwakeup() from an MP-protected taskqueue.
Before, we would call selwakeup() from interrupt context.  This
was working as long kern.intr_mpsafe == 0.  selwakeup() however needs
the MP lock held, so we can't call it directly from the interrupt
when running with kern.intr_mpsafe=1.

Instead, perform the wakeup from a "bottom half" taskqueue SWI to make the
sound devices intr_mpsafe, as they claim to be.