HAMMER 16B/many: Fix data overwrite case.
HAMMER will often write out data records for an inode before writing out
the newly created inode itself.  This is desireable because it allows
HAMMER to avoid replacing the on-disk inode record every time the file
size changes.

Fix a bug related to this issue where HAMMER was not checking for on-disk
data records in the data overwrite case prior to the inode being written
out.