HAMMER 31A/many: File data size optimization
I've been meaning to do this for a while.  When writing a data record for
the last block in a file, for example for small files, there is no need
to write out the entire 16K buffer.  Just write out the actual number of
bytes of data to file EOF.

This results in *EXTREME* data compaction and very high performance when
manipulating small files.  The smallest possible file requires two 96 byte
records (inode record and data record), two 64 byte B-Tree elements,
a 64 byte data element (the file data), and a 128 byte data element
(the inode structure), for a grand total of 512 bytes.