Add an experimental driver for NICs using Silan Microelectronics' SC92301 chip, some of which seem to be known as Rsltek [sic] 8139D. This is a port of Silan's own FreeBSD 4.7 driver which was written by one 'gaoyonghong'.
It's up to the point where it works with the "Noganet KN-8139D" product,
but it still gives occasional errors/warnings on the console. Also, some
areas need to be brought more up to date.

Therefore, the if_sln.ko module is built, but the driver is not yet in
GENERIC.

Tested-by: Damian Vicino <dvicino@dc.uba.ar>