Sync with NetBSD:
Use the service name to getaddrinfo() (along with the host name), so that
features such as DNS Service Discovery have a better chance of working.

Display the service name in various status & error messages.

Don't getservbyname() the :port component of a URL; RFC 3986 says it's
just an unsigned number, not a service name.

Only attempt to el_parse() a command unknown by the default parser
if editing is enabled.
Fixes NetBSD PR bin/38589

Add epsv6 and epsv to disable extended passive mode for ipv6 or both ipv4
and ipv6 respectively.