Add the UF_NOHISTORY and SF_NOHISTORY chflags flags.  The nohistory flag allows you to mount a HAMMER filesystem normally and still specify that certain files and directory not retain historical information.
If set on a directory the flag will be inherited by any objects
created within that directory.

Adjust UFS to inherit the NOHISTORY and NODUMP flags from the parent
directory when creating new objects.  NODUMP used to be ufs-specific
and UFS's backup/restore remembered the inheritance.  As a more generic
flag it needs to be inherited within the filesystem.  Note that UFS
has no historical retention capability and ignores the NOHISTORY flag,
but we might use it with the journaling audit trail later.