Use humanize_number() from libtool instead of rolling our own version for -h and -H.
As a side effect, this fixes an issue with showing a negative value in
the 'Avail' column (for UFS file systems which are using the MINFREE
reserve) by casting f_bavail to signed when passing it to prthumanval(),
like it was done before already for fstoblk().

f_bavail used to be signed in struct fsstat but is now unsigned in struct
vfsstat (in accordance with SUS).

Taken-from: FreeBSD