HAMMER: Mass storage flush command support
Add mass storage flush command support to HAMMER.  The HAMMER flush cycle
issues async I/O in parallel and waits for it to complete before moving
onto the next stage.  This occurs in two places.  The new flush command
is issued after the async I/O, and then we wait on everything.

Due to HAMMER's ability to gang writes asynchronously between stages the
addition of the flush should not significantly impact performance.