Add CRAM-MD5 authentication support for the DragonFly Mail Agent. This is the first piece of Max's work for the Google Summer of Code 2008.  All other new features will follow after evaluation/review :)
Besides the CRAM code there is new code within base64.c (BSD licensed and
from the University of Stockholm) and within crypto.c derived from RFC 2104.

Note: This code is tested and works.  If you find a bug, please report back
to the bugs@ list.

Thanks a lot for the good work Max.

Submitted-by: Max Lindner <gisanka@googlemail.com>
Sponsored-by: Google Summer of Code 2008