Remove the -1 option from the installer.
This option is specific to the old, now removed, ISC dhcp client.  The new
OpenBSD dhcp client does not support this option.  If a user wants to set an
IP address with the installer prior to this commit, it fails and the user sees
an error message.

The meaning of the -1 option was: "Try to get a lease only once.  If it fails,
quit."  Due to this commit, this behaviour changes.  I think its better to have
an installer which tries to get a lease more than once that having an installer
which does not try to get a lease at all.