Fix a double-flush which was occuring for every unlinked inode, resulting in very inefficient flushing operations which also prevented any inode from being reclaimed until the entire flush batch (which could be upwards of 40,000 inodes) completed.
The double-flush was caused by the HAMMER_INODE_DELETING flag being
improperly set twice.

Clear the HAMMER_INODE_REFLUSH flag if a reflush was queued but the
modmask is cleared by the prior flush.

Adjust the directory mtime before flushing an underlying inode, avoiding
a double flush of the directory inode.

Reported-by: Hasso Tepper <hasso@estpak.ee>