Restore the semantic of callout_active() testing on tcp timers.
Originally there is no time gap between the running of the tcp timer
handler and the deactivation of the tcp timer callout, but the message
based tcp timer has a time gap in between these two actions.  This
time gap affects the code path which depends on the current state of
the tcp timer, i.e. return value of callout_active(tcp_timer).  To
close this time gap, we take the pending and running tcp timer tasks
into consideration when testing the current state of the tcp timer.

Reviewed-by: dillon@