This is a major revamping of the pageout and low-memory handling code.
The pageout daemon now detects out-of-memory conditions and properly
kills the largest process(es).  This condition occurs when swap is
full (or you have no swap) and most of the remaining VM pages in memory
have become dirty.  With no swap to page to the dirty pages squeeze out
the clean ones.  The pageout daemon detects the case and starts killing
processes.

The pageout daemon now detects stress in the form of excess cpu use
and tries to reduce its cpu footprint when that occurs.  Excess cpu use
can occur when the only pages left in-core are dirty and there is nowhere
to swap them to.  Previously if this case occured the system would basically
just stop working.

These changes make the system truely have VM = RAM+SWAP.  If you 1G of ram
and 1G of swap the system can run up to 2G worth of processes.