Add nsswitch support.
The nsswitch.conf(5) manual page has a description of nsswitch.
Curiously, we already had this manual page, even though we didn't
support it.

/etc/host.conf is removed from src/, but if host.conf exists and
nsswitch.conf does not, nsswitch.conf will be created using its
contents.

Included in this commit is a framework for nsswitch caching, nscd(8),
but it relies on a few upcoming changes to our libc before it will work.
For now, it's turned off.

Also this commit includes hesiod support which is not compiled by
default.  Add WANT_HESIOD=true to make.conf to get it working.

Obtained-from: FreeBSD