Make the "Configure installed system" functionality work
This button did not work, because the installer could not determine whether
the installed fs is HAMMER or UFS.  Ask the user before mounting the partition.
I chose the simple dialog rather than a automatic solution because it would
make the installer even more ugly (i.e. it would results in some grep+awk+foobar
magic).